Small fixes to make PSR2 tests pass with php-cs-fixer
